--- maibaduoduo-microservice/maibaduoduo-business-master/program-master/program-core/src/main/java/com/maibaduoduo/analysis/item/dao/ProgramItemDao.java
@@ -1,24 +1,0 @@
-/**
- * Copyright (c) 2019-2023 SAAS开源 All rights reserved.
- *
- * SAAS系统设计研发交流
- *
- * https://www.maibaduoduo.com
- */
-package com.maibaduoduo.analysis.item.dao;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.maibaduoduo.analysis.item.entity.ProgramItemEntity;
-import org.apache.ibatis.annotations.Mapper;
-
-/**
- * 
- * 
- * @author saas
- * @email lifengwork@yeah.net
- * @date 2023-05-06 20:24:36
- */
-@Mapper
-public interface ProgramItemDao extends BaseMapper<ProgramItemEntity> {
-	
-}
