+++ maibaduoduo-microservice/maibaduoduo-business-master/program-master/program-core/src/main/java/com/maibaduoduo/program/entity/ProgramEntity.java
@@ -1,0 +1,59 @@
+/**
+ * Copyright (c) 2019-2023 SAAS开源 All rights reserved.
+ *
+ * SAAS系统设计研发交流
+ *
+ * https://www.maibaduoduo.com
+ */
+package com.maibaduoduo.program.entity;
+
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
+
+import java.io.Serializable;
+import java.util.Date;
+import lombok.Data;
+
+/**
+ * 
+ * 
+ * @author saas
+ * @email lifengwork@yeah.net
+ * @date 2023-05-06 20:24:36
+ */
+@Data
+@TableName("saas_program")
+public class ProgramEntity implements Serializable {
+	private static final long serialVersionUID = 1L;
+
+	/**
+	 * 
+	 */
+	@TableId
+	private Long id;
+	/**
+	 * 
+	 */
+	private String programName;
+	/**
+	 * 
+	 */
+	private String programDesc;
+	/**
+	 * 
+	 */
+	private String programStatus;
+	/**
+	 * 创建者ID
+	 */
+	private Long createUserId;
+	/**
+	 * 创建时间
+	 */
+	private Date createTime;
+	/**
+	 * 租户编码
+	 */
+	private String tenantId;
+
+}
