+++ maibaduoduo-microservice/maibaduoduo-business-master/program-master/program-core/src/main/resources/mapper/program/ProgramDao.xml
@@ -6,10 +6,10 @@
 -->
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.maibaduoduo.analysis.dao.ProgramDao">
+<mapper namespace="com.maibaduoduo.program.dao.ProgramDao">
 
     <!-- 映射 -->
-    <resultMap type="com.maibaduoduo.analysis.entity.ProgramEntity" id="programMap">
+    <resultMap type="com.maibaduoduo.program.entity.ProgramEntity" id="programMap">
                     <result property="id" column="id"/>
                     <result property="programName" column="program_name"/>
                     <result property="programDesc" column="program_desc"/>
