--- maibaduoduo-microservice/maibaduoduo-business-master/purchase-master/purchase-facade/src/main/java/com/maibaduoduo/analysis/facade/api/PurchaseApiFallbackFactory.java
@@ -1,24 +1,0 @@
-/**
- * Copyright (c) 2019-2023 SAAS开源 All rights reserved.
- *
- * SAAS系统设计研发交流
- *
- * https://www.maibaduoduo.com
- */
-package com.maibaduoduo.analysis.facade.api;
-
-import com.maibaduoduo.api.ApiFallbackFactory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.Primary;
-import org.springframework.stereotype.Component;
-
-@Component
-@Primary
-public class PurchaseApiFallbackFactory extends ApiFallbackFactory {
-    Logger logger = LoggerFactory.getLogger(this.getClass());
-    @Override
-    public PurchaseFacade fallbackFactory(Throwable throwable) {
-        return null;
-    }
-}
