--- maibaduoduo-microservice/maibaduoduo-business-master/purchase-master/purchase-facade/src/main/java/com/maibaduoduo/analysis/facade/api/PurchaseFacade.java
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) 2019-2023 SAAS开源 All rights reserved.
- *
- * SAAS系统设计研发交流
- *
- * https://www.maibaduoduo.com
- */
-package com.maibaduoduo.analysis.facade.api;
-
-import com.maibaduoduo.api.ApiFacade;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.stereotype.Component;
-
-
-@Component
-@FeignClient(name = "program-service",configuration = {PurchaseRestTemplateConfiguration.class},fallbackFactory = PurchaseApiFallbackFactory.class)
-public interface PurchaseFacade extends ApiFacade {
-
-}
