--- maibaduoduo-microservice/maibaduoduo-business-master/supplier-master/supplier-facade/src/main/java/com/maibaduoduo/analysis/facade/api/SupplierFacade.java
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) 2019-2023 SAAS开源 All rights reserved.
- *
- * SAAS系统设计研发交流
- *
- * https://www.maibaduoduo.com
- */
-package com.maibaduoduo.analysis.facade.api;
-
-import com.maibaduoduo.api.ApiFacade;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.stereotype.Component;
-
-
-@Component
-@FeignClient(name = "program-service",configuration = {SupplierRestTemplateConfiguration.class},fallbackFactory = SupplierApiFallbackFactory.class)
-public interface SupplierFacade extends ApiFacade {
-
-}
